split things up/organize
